# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"AI-AI Stroke buddy.","description":"Poststroke AI Assistant.","version":"0.1.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Get auth token","description":"Return access and refresh tokens.","operationId":"auth_user_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Auth"],"summary":"Refresh auth token","description":"Rotate refresh token and issue new access token.","operationId":"refresh_token_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["Auth"],"summary":"Log user out","description":"Revoke refresh token.","operationId":"logout_user_api_v1_auth_logout_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Logout User Api V1 Auth Logout Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer":[]}]}},"/api/v1/users/register":{"post":{"tags":["Users"],"summary":"New user creation","description":"User creation endpoint.","operationId":"create_user_api_v1_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["Users"],"summary":"Get current user info","description":"Get current user info.","operationId":"get_your_user_info_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}}},"security":[{"Bearer":[]}]},"delete":{"tags":["Users"],"summary":"Deactivate current user","description":"Soft delete — sets is_active to False.","operationId":"deactivate_user_api_v1_users_me_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"Bearer":[]}]},"patch":{"tags":["Users"],"summary":"Update user info","description":"Update current user info.","operationId":"update_user_info_api_v1_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer":[]}]}},"/api/v1/users/{user_id}/assign-doctor":{"patch":{"tags":["Users"],"summary":"Assign or unassign a doctor to a patient","description":"Assign or unassign a doctor to a patient. Privileged endpoint.","operationId":"assign_doctor_api_v1_users__user_id__assign_doctor_patch","security":[{"Bearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignDoctorUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me/timezone":{"patch":{"tags":["Users"],"summary":"Sync device timezone","description":"Call on app load to keep timezone in sync.","operationId":"sync_timezone_api_v1_users_me_timezone_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimezoneUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer":[]}]}},"/api/v1/diary":{"get":{"tags":["Diary"],"summary":"Get diary entries","description":"Return paginated diary entries for the current user.","operationId":"get_diary_entries_api_v1_diary_get","security":[{"Bearer":[]}],"parameters":[{"name":"entry_type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/EntryType"},{"type":"null"}],"title":"Entry Type"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DiaryEntryInfo"},"title":"Response Get Diary Entries Api V1 Diary Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/diary/{entry_id}":{"get":{"tags":["Diary"],"summary":"Get a single diary entry","description":"Return a single diary entry belonging to the current user.","operationId":"get_diary_entry_api_v1_diary__entry_id__get","security":[{"Bearer":[]}],"parameters":[{"name":"entry_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Entry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiaryEntryInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Diary"],"summary":"Delete a diary entry","description":"Delete a diary entry belonging to the current user.","operationId":"delete_diary_entry_api_v1_diary__entry_id__delete","security":[{"Bearer":[]}],"parameters":[{"name":"entry_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Entry Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/stream":{"post":{"tags":["Chat"],"summary":"Chat Stream","description":"Create main chat endpoint.","operationId":"chat_stream_api_v1_chat_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer":[]}]}},"/api/v1/reminders/":{"get":{"tags":["Reminders"],"summary":"Get Reminders","description":"Get reminders.","operationId":"get_reminders_api_v1_reminders__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ReminderOut"},"type":"array","title":"Response Get Reminders Api V1 Reminders  Get"}}}}},"security":[{"Bearer":[]}]}},"/api/v1/reminders/{reminder_id}":{"delete":{"tags":["Reminders"],"summary":"Delete Reminder","description":"Delete reminder.","operationId":"delete_reminder_api_v1_reminders__reminder_id__delete","security":[{"Bearer":[]}],"parameters":[{"name":"reminder_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Reminder Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reminders/vapid-public-key":{"get":{"tags":["Reminders"],"summary":"Get Vapid Public Key","description":"Frontend needs this to subscribe to push notifications.","operationId":"get_vapid_public_key_api_v1_reminders_vapid_public_key_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VapidPublicKeyOut"}}}}}}},"/api/v1/reminders/push-subscription":{"post":{"tags":["Reminders"],"summary":"Save Push Subscription","description":"Save or update browser push subscription.","operationId":"save_push_subscription_api_v1_reminders_push_subscription_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PushSubscriptionIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Save Push Subscription Api V1 Reminders Push Subscription Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer":[]}]},"delete":{"tags":["Reminders"],"summary":"Delete Push Subscription","description":"Call when user disables notifications.","operationId":"delete_push_subscription_api_v1_reminders_push_subscription_delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PushSubscriptionIn"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer":[]}]}}},"components":{"schemas":{"AssignDoctorUpdate":{"properties":{"doctor_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doctor Id"}},"additionalProperties":false,"type":"object","title":"AssignDoctorUpdate","description":"Schema for assigning or unassigning a doctor to a patient."},"AuthData":{"properties":{"login":{"type":"string","format":"phone","title":"Login"},"password":{"type":"string","format":"password","title":"Password","writeOnly":true}},"additionalProperties":false,"type":"object","required":["login","password"],"title":"AuthData","description":"Auth scheme."},"AuthToken":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type"}},"additionalProperties":false,"type":"object","required":["access_token","refresh_token","token_type"],"title":"AuthToken","description":"Token scheme."},"ChatRequest":{"properties":{"session_id":{"type":"string","format":"uuid","title":"Session Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["session_id","message"],"title":"ChatRequest","description":"Chat request schema."},"DiaryEntryInfo":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"entry_type":{"$ref":"#/components/schemas/EntryType"},"entry_json":{"additionalProperties":true,"type":"object","title":"Entry Json"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","user_id","entry_type","entry_json","created_at","updated_at"],"title":"DiaryEntryInfo","description":"Diary entry response schema."},"EntryType":{"type":"string","enum":["blood_pressure","blood_test","medication","symptom"],"title":"EntryType","description":"Type of diary entry."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Medication":{"properties":{"name":{"type":"string","title":"Name"},"dose_mg":{"type":"integer","title":"Dose Mg"},"frequency":{"type":"string","title":"Frequency"}},"additionalProperties":false,"type":"object","required":["name","dose_mg","frequency"],"title":"Medication","description":"Medication schema."},"PushSubscriptionIn":{"properties":{"endpoint":{"type":"string","title":"Endpoint"},"p256dh":{"type":"string","title":"P256Dh"},"auth":{"type":"string","title":"Auth"}},"type":"object","required":["endpoint","p256dh","auth"],"title":"PushSubscriptionIn","description":"Push sub schema."},"RefreshTokenRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"additionalProperties":false,"type":"object","required":["refresh_token"],"title":"RefreshTokenRequest","description":"Refresh token scheme."},"ReminderOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"reminder_type":{"type":"string","title":"Reminder Type"},"med_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Med Name"},"time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Time"},"days":{"items":{},"type":"array","title":"Days"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","reminder_type","med_name","time","days","is_active","created_at"],"title":"ReminderOut","description":"Reminder schema."},"Role":{"type":"string","enum":["patient","relative","doctor","admin"],"title":"Role","description":"User role."},"Sex":{"type":"string","enum":["male","female","unknown"],"title":"Sex","description":"Patient sex."},"StrokeHemSubType":{"type":"string","enum":["intracerebral hemorrhage","subarachnoid hemorrhage"],"title":"StrokeHemSubType","description":"Hemorrhagic stroke subtype."},"StrokeTOASTSubType":{"type":"string","enum":["large-artery atherosclerosis","cardioembolism","small-vessel occlusion","stroke of other determined etiology","stroke of undetermined etiology"],"title":"StrokeTOASTSubType","description":"Ischenic stroke subtype."},"StrokeType":{"type":"string","enum":["ischemic","hemorrhagic","unknown"],"title":"StrokeType","description":"Type of stroke."},"TimezoneUpdate":{"properties":{"timezone":{"type":"string","title":"Timezone"}},"additionalProperties":false,"type":"object","required":["timezone"],"title":"TimezoneUpdate","description":"Schema for timezone sync."},"UserCreate":{"properties":{"phone":{"type":"string","format":"phone","title":"Phone"},"password":{"type":"string","format":"password","title":"Password","writeOnly":true}},"additionalProperties":false,"type":"object","required":["phone","password"],"title":"UserCreate","description":"Pydantic-schema for user creation."},"UserInfo":{"properties":{"phone":{"type":"string","format":"phone","title":"Phone"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"sex":{"anyOf":[{"$ref":"#/components/schemas/Sex"},{"type":"null"}]},"stroke_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Stroke Date"},"recurrent_stroke":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Recurrent Stroke"},"stroke_type":{"anyOf":[{"$ref":"#/components/schemas/StrokeType"},{"type":"null"}]},"stroke_toast_subtype":{"anyOf":[{"$ref":"#/components/schemas/StrokeTOASTSubType"},{"type":"null"}]},"stroke_hemo_subtype":{"anyOf":[{"$ref":"#/components/schemas/StrokeHemSubType"},{"type":"null"}]},"medications":{"anyOf":[{"items":{"$ref":"#/components/schemas/Medication"},"type":"array"},{"type":"null"}],"title":"Medications"},"doctor_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doctor Id"},"daily_checkin_enabled":{"type":"boolean","title":"Daily Checkin Enabled","default":true},"timezone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timezone"}},"additionalProperties":false,"type":"object","required":["phone"],"title":"UserInfo","description":"Pydantic-schema for user info show."},"UserUpdate":{"properties":{"phone":{"anyOf":[{"type":"string","format":"phone"},{"type":"null"}],"title":"Phone"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"sex":{"anyOf":[{"$ref":"#/components/schemas/Sex"},{"type":"null"}]},"stroke_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Stroke Date"},"recurrent_stroke":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Recurrent Stroke"},"stroke_type":{"anyOf":[{"$ref":"#/components/schemas/StrokeType"},{"type":"null"}]},"stroke_toast_subtype":{"anyOf":[{"$ref":"#/components/schemas/StrokeTOASTSubType"},{"type":"null"}]},"stroke_hemo_subtype":{"anyOf":[{"$ref":"#/components/schemas/StrokeHemSubType"},{"type":"null"}]},"role":{"anyOf":[{"$ref":"#/components/schemas/Role"},{"type":"null"}]},"daily_checkin_enabled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Daily Checkin Enabled"},"timezone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timezone"}},"additionalProperties":false,"type":"object","title":"UserUpdate","description":"Pydantic-schema for user info update."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VapidPublicKeyOut":{"properties":{"public_key":{"type":"string","title":"Public Key"}},"type":"object","required":["public_key"],"title":"VapidPublicKeyOut","description":"VapidPubKey schema."}},"securitySchemes":{"Bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}